fix: Fix errors in P,Q datatype

+++ XOR/synapseArray5by5.h
@@ -61,10 +61,10 @@ public:
     // (SGD) Stochastic Gradient Descent
     int p[5];
     int q[5];
-    int P1[5];
-    int Q1[5];
-    int P2[5];
-    int Q2[5];
+    double P1[5];
+    double Q1[5];
+    double P2[5];
+    double Q2[5];
     /*---------------------methods----------------*/
 
 public:
